+++ ImageFeed/ViewConrollers/ImagesListView/ImagesListViewController.swift
@@ -58,12 +58,8 @@ extension ImagesListViewController: UITableViewDataSource {
         return photosName.count
     }
     
-    
-    
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
         let cell = tableView.dequeueReusableCell(withIdentifier: ImagesListCell.reuseIdentifier, for: indexPath)
-        
         
         guard let imageListCell = cell as? ImagesListCell else {
             return UITableViewCell()
